pyproject.toml
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "idp-fuel-tracker"
description = ""
version = "0.1.0"
authors = [{name = "Josh Peak", email = "neozenith.dev@gmail.com"}]
readme = "README.md"
license = {text = "MIT License"}
requires-python = ">=3.10"

dependencies = [
    "pillow-heif",
    "boto3",
    "amazon-textract-response-parser",
    "amazon-textract-overlayer",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit",
    "ruff",
    "isort",
    "md-toc"
]
